backend_new/requirements.txt
fastapi
uvicorn
pandas
google-genai
python-dotenv
sentence-transformers
scikit-learn
mysql-connector-python
pymysql
firebase-admin
sqlalchemy
pydantic
websockets
